rpc framework test
