+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,2 @@
+Manifest-Version: 1.0
+Main-Class: com.mytest.rpc.service.publish.RpcFrameworkBoot
