+++ src/test/java/com/mytest/rpc/RpcFrameworkTest.java
@@ -4,16 +4,14 @@ import java.io.IOException;
 
 import org.junit.Test;
 
+import com.mytest.rpc.service.HelloService;
+import com.mytest.rpc.service.publish.RpcFrameworkBoot;
+
 /**
  * @author liqingyu
  * @since 2018/09/03
  */
 public class RpcFrameworkTest {
-
-    public void publish() throws IOException {
-        HelloService helloService = new HelloServiceImpl();
-        RpcFramework.publish(helloService, 8888);
-    }
 
 
     @Test
@@ -22,26 +20,21 @@ public class RpcFrameworkTest {
             @Override
             public void run() {
                 try {
-                    new RpcFrameworkTest().publish();
+                    String[] args = new String[]{};
+                    new RpcFrameworkBoot().main(args);
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
             }
         }).start();
+        try {
+            //等待服务器发布服务
+            Thread.sleep(300);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
         HelloService service = RpcFramework.call(HelloService.class, "127.0.0.1", 8888);
         String result = service.hello("liqingyu");
         System.out.println(result);
-    }
-
-    public interface HelloService {
-        String hello(String name);
-    }
-
-    public class HelloServiceImpl implements HelloService {
-
-        @Override
-        public String hello(String name) {
-            return "RPC test, hello " + name;
-        }
     }
 }
